Keep the account route under the client file-size ratchet.
Move email-change and former-address client state into a factory so
account.tsx stays under the 800-line budget after the claims UI.
